test: split test_event_drafts to satisfy 500-line limit (Issue 721)
The 16 lines added in this PR pushed test_event_drafts.py past the
500-line hard limit. Move the shared draft fixtures into conftest.py
(single source of truth) and extract the publish/delete lifecycle tests
into test_event_publish.py.
